requirements.txt
python-dotenv == 1.0.1
langchain == 0.2.5
langchain-cli == 0.0.25
langchain-openai == 0.1.8
langchain-community == 0.2.5
langchain-chroma == 0.1.1
langchainhub == 0.1.20
streamlit == 1.35.0
azure-storage-blob == 12.21.0
pypdf == 4.3.1
clipboard == 0.0.4